Switch to CodeRay
Because CodeRay is pure Ruby, it fits more nicely into a Ruby project
than Pygments (not to disparage pygments.rb, but getting that to work
can be annoying).

This commit uses the default CSS for CodeRay, generated by the command
coderay stylesheet > content/css/coderay.css
It's pretty ugly, to be honest. I'm going to add a better one soon.