Add HTMLEscape helper
I didn't need any of the methods in the Tagging helper, but I did need
the methods in this helper (which was included by the Tagging helper).
Oops. Because I was running nanoc through guard, maybe it didn't unload
the helper, so the site continued to compile correctly despite having
removed the helper from helpers.rb.