Move year/month/day index generation to lib
Another step of the preprocess clean up. This commit moves the logic for
year/month/day indices into lib, and cleans it up a lot.

The trick used in the previous commit (where a list is passed into a
partial layout via an item attribute) is also used here, to avoid re-
computing the list of articles in a given year/month/day. This probably
saves time compared to repeatedly select-ing on the articles array, but
I haven't done the benchmarks.