Use Date.strftime for year/month/day indices
This commit improves the generation of item identifiers and titles for
the year/month/day indices. Instead of hacking them by hand, we can use
Date.strftime, so that the formats are easier to change going forward.

Formats for the titles of year and month indices have been added to
nanoc.yaml to accompany this change.