Rename Pygments CSS
Seeing as we have two syntax highlighters living under the same roof,
it's a bit vague for the Pygments CSS to be called "syntax.css".

I didn't want to rename it in the previous commit, because then the
history of the file might have been disconnected. This way, we should
be able to track the history of this file with git log --follow.

I've included both CSS files in the header. I have tested (sparingly)
and they don't seem to interfere with each other.