Reorganize layouts
This commit creates the layouts/includes and layouts/partials subfolders
for organizing layouts. I wanted to do this for a while, but I didn't
know if nanoc would be able to handle the subfolders in layout names
until I tried it.

The includes folder is mostly analogous to the Jekyll _includes folder.
It's for layouts that are rendered as-is inside other layouts.

The partials folder is for layouts that get rendered with certain
attributes, to serve as part of another page. This folder is also used
for layouts that can be used for dynamically generated items, in the
preprocess block.